+++ chevah/compat/testing/filesystem.py
@@ -96,13 +96,15 @@ class LocalTestFilesystem(LocalFilesystem):
         segments = self.temp_segments
         return self.getRealPathFromSegments(segments)
 
-    def createFile(self, segments, length=0, access_time=None, content=None):
+    def createFile(
+        self, segments, length=0, access_time=None, content=None, mode=0o666,
+            ):
         '''Creates a file.
 
         Raise AssertionError if file already exists or it can not be created.
         '''
         assert not self.isFile(segments), 'File already exists.'
-        new_file = self.openFileForWriting(segments)
+        new_file = self.openFileForWriting(segments, mode=mode)
         if content is None:
             value = 'a'
             content = ''
@@ -138,8 +140,8 @@ class LocalTestFilesystem(LocalFilesystem):
         """
         Write content into file replacing existing content.
         """
-        opened_file = self.openFileForWriting(segments, utf8=True)
-        opened_file.write(content)
+        opened_file = self.openFileForWriting(segments)
+        opened_file.write(content.encode('utf-8'))
         opened_file.close()
 
     def _makeFilename(self, prefix=u'', suffix=u''):
@@ -161,7 +163,7 @@ class LocalTestFilesystem(LocalFilesystem):
         path = os.path.join(self.temp_path, name)
         return (path, segments)
 
-    def pathInTemp(self, cleanup):
+    def pathInTemp(self, cleanup, prefix=u'', suffix=u''):
         """
         Return a path and segments pointing to a temp location, which will be
         cleaned up.
@@ -364,26 +366,14 @@ class LocalTestFilesystem(LocalFilesystem):
 
         By default, the content is returned as Unicode.
         """
-        opened_file = self.openFileForReading(segments, utf8=utf8)
+        opened_file = self.openFileForReading(segments)
         result = opened_file.read()
         opened_file.close()
+
+        if utf8:
+            return result.decode('utf-8')
+
         return result
-
-    def getFileLines(self, segments, utf8=True):
-        """
-        Return a list with all lines from file.
-
-        By default, the content is returned as Unicode.
-        """
-        opened_file = self.openFileForReading(segments, utf8=utf8)
-        content = []
-        try:
-            for line in opened_file:
-                content.append(line.rstrip())
-        finally:
-            opened_file.close()
-
-        return content
 
     def replaceFileContent(self, segments, rules):
         """
@@ -392,7 +382,7 @@ class LocalTestFilesystem(LocalFilesystem):
         It takes a list for tuples [(pattern1, substitution1), (pat2, sub2)]
         and applies them in order.
         """
-        opened_file = self.openFileForReading(segments, utf8=True)
+        opened_file = self.openFileForReading(segments)
         altered_lines = []
         for line in opened_file:
             new_line = line
@@ -406,7 +396,7 @@ class LocalTestFilesystem(LocalFilesystem):
             altered_lines.append(new_line)
         opened_file.close()
 
-        opened_file = self.openFileForWriting(segments, utf8=True)
+        opened_file = self.openFileForWriting(segments)
         for line in altered_lines:
             opened_file.write(line)
         opened_file.close()